rename (global) : Redis 관련 application 계층 파일 각 사용도메인 내로 이동
