+++ src/main/java/com/backend/immilog/global/infrastructure/persistence/config/QueryDSLConfig.java
@@ -1,4 +1,4 @@
-package com.backend.immilog.global.persistence;
+package com.backend.immilog.global.infrastructure.persistence.config;
 
 import com.querydsl.jpa.JPQLTemplates;
 import com.querydsl.jpa.impl.JPAQueryFactory;
